+++ CHANGES.md
@@ -12,7 +12,11 @@
 ### Features
 
 - Add a migration framework. Automatic migrations can now be called using
-  `batou migrate`.
+  `./batou migrate`.
+  ([#185](https://github.com/flyingcircusio/batou/issues/185))
+
+- Automatically migrate environments and secrets to the new structure using
+  `./batou migrate`.
   ([#185](https://github.com/flyingcircusio/batou/issues/185))